requirements.txt
streamlit
pandas
folium
streamlit-folium
pytest
requests
python-dotenvstreamlit
requests
python-dotenv
pytest
folium
streamlit-folium
